refactored modules in psycopg_async_listen
